patch: increase Leader GetGTID timeout
When mysqld is under high pressure, the value of timeout that the leader execute GetGTID each time should be different.

The value of timeout needs to increase as the number of raft vote count. This case has no effect on the leader when the pressure is normal.

If raft role is follower or candidate, it's no need to change the timeout.